Update BgApi.cs
Fix for Issue #7.
It is not necessary to reset the device on every Open, depending on how you workd on it. And it takes a lot of time (5 -6 Seconds) when more than one stick is present, because of the windows serial port Close-Open Issue.
So I made the reset optional. It's backward-compatible, so that existing code should work the same way as before.